address
- vilgotf: combine action structs
- vilgotf: event type flags
